+++ src/events_impl/mod.rs
@@ -1,12 +1,14 @@
+pub use core::*;
+
 pub use adapt::*;
 pub use bhv_ext::BhvExt;
 pub use composite::*;
-pub use core::*;
+pub use decor::*;
 pub use events::*;
 
 mod adapt;
+mod bhv_ext;
 mod composite;
 mod core;
 mod decor;
 mod events;
-mod bhv_ext;