+++ src/old_impl/adapt.rs
@@ -14,6 +14,11 @@ pub struct Action<Ctx, A>(A, PhantomData<Ctx>)
     where
         A: FnMut(&mut Ctx);
 
+/// The type of the result of [`async_action`].
+#[derive(Clone)]
+pub struct AsyncAction<Ctx, A>(A, PhantomData<Ctx>)
+    where A: FnMut(&mut Ctx) -> Status;
+
 /// Adapt a predicate into a behavior, returning [`Status::Success`] if
 /// the predicate returns `true` and [`Status::Failure`] otherwise.
 ///
@@ -54,12 +59,55 @@ pub fn action<Ctx, A>(a: A) -> Action<Ctx, A>
     Action(a, PhantomData)
 }
 
+/// Wrap a function returning a [`Status`] into a behavior.
+///
+/// # Example
+///
+/// ```
+/// use bhv::*;
+///
+/// struct State {
+///     a: i32,
+///     b: i32,
+///     n: i32,
+/// }
+///
+/// impl State {
+///     fn new(steps: i32) -> Self {
+///         Self { a: 0, b: 1, n: steps }
+///     }
+/// }
+///
+/// let mut state = State::new(10); // show first 10 fibonacci numbers
+///
+/// let code = async_action(|s: &mut State| {
+///     if s.n == 0 {
+///         Status::Success
+///     } else {
+///         print!("{} ", s.b);
+///         let tmp = s.b;
+///         s.b += s.a;
+///         s.a = tmp;
+///
+///         s.n -= 1;
+///         Status::Running
+///     }
+/// });
+///
+/// code.execute(&mut state);
+///
+/// assert_eq!(state.b, 89);
+/// ```
+#[inline]
+pub fn async_action<Ctx, A>(a: A) -> AsyncAction<Ctx, A>
+    where A: FnMut(&mut Ctx) -> Status { AsyncAction(a, PhantomData) }
+
 impl<Ctx, C> Bhv for Cond<Ctx, C>
     where
         C: Fn(&Ctx) -> bool,
 {
     type Context = Ctx;
-
+    #[inline]
     fn update(&mut self, ctx: &mut Self::Context) -> Status {
         if self.0(ctx) {
             Status::Success
@@ -74,9 +122,19 @@ impl<Ctx, A> Bhv for Action<Ctx, A>
         A: FnMut(&mut Ctx),
 {
     type Context = Ctx;
-
+    #[inline]
     fn update(&mut self, ctx: &mut Self::Context) -> Status {
         self.0(ctx);
         Status::Success
+    }
+}
+
+impl<Ctx, A> Bhv for AsyncAction<Ctx, A>
+    where A: FnMut(&mut Ctx) -> Status,
+{
+    type Context = Ctx;
+    #[inline]
+    fn update(&mut self, ctx: &mut Self::Context) -> Status {
+        self.0(ctx)
     }
 }