+++ src/old_impl/mod.rs
@@ -1,12 +1,13 @@
+pub use self::adapt::*;
+pub use self::async_composite::*;
+pub use self::bhv_ext::BhvExt;
+pub use self::composite::*;
+pub use self::core::*;
+pub use self::decor::*;
+
 mod adapt;
 mod bhv_ext;
 mod composite;
 mod core;
 mod decor;
-
-pub use self::core::*;
-pub use self::adapt::*;
-pub use self::composite::*;
-pub use self::decor::*;
-
-pub use self::bhv_ext::BhvExt;
+mod async_composite;