+++ CHANGELOG.md
@@ -6,6 +6,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2024-03-03
+
+### Added
+
+- `async_action` that wraps a function returning a `Status` into a behavior.
+- `RunIf` node on the old implementation that runs a node only if a predicate is fulfilled.
+- `WhenAny`/`WhenAll` on the old implementation, that run a set of nodes until the first/all of them complete
+successfully.
+
+
 ## [0.3.0] - 2024-02-24
 
 ### Added